Fix
add int16_t in calculation part.

+++ examples/d6t-1a/d6t-1a.ino
@@ -61,12 +61,12 @@ bool judge_seatOccupancy(void) {
   }
   if (totalCount > comparingNumInc){    
     if (occuPix == false){
-      if (seqData[0] - seqData[comparingNumInc] > threshHoldInc){
+      if ((int16_t)(seqData[0] - seqData[comparingNumInc]) > (int16_t)threshHoldInc){
         occuPix = true;
       }
     }
     else{   //resultOccupancy == true
-      if (seqData[comparingNumDec] - seqData[0] > threshHoldDec){
+      if ((int16_t)(seqData[comparingNumDec] - seqData[0]) > (int16_t)threshHoldDec){
         occuPix = false;
       }
     }
